requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
pylint==2.5.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.15.0
toml==0.10.1
tweepy==3.10.0
urllib3==1.26.2
wrapt==1.12.1
